include products queries and models
